requirements.txt
boto3
mysql-connector-python
requests
